+++ minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/item/tooltip/StateWriter.java
@@ -1,10 +1,13 @@
 package com.sigmundgranaas.forgero.minecraft.common.item.tooltip;
 
+import com.sigmundgranaas.forgero.core.property.v2.cache.ContainsFeatureCache;
+import com.sigmundgranaas.forgero.core.property.v2.cache.PropertyTargetCacheKey;
 import com.sigmundgranaas.forgero.core.soul.SoulContainer;
 import com.sigmundgranaas.forgero.core.state.Composite;
 import com.sigmundgranaas.forgero.core.state.LeveledState;
 import com.sigmundgranaas.forgero.core.state.State;
 import com.sigmundgranaas.forgero.minecraft.common.item.tooltip.writer.*;
+import com.sigmundgranaas.forgero.minecraft.common.toolhandler.UnbreakableHandler;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.text.MutableText;
 import net.minecraft.text.Text;
@@ -12,6 +15,7 @@ import net.minecraft.util.Formatting;
 
 import java.util.List;
 
+import static com.sigmundgranaas.forgero.core.condition.Conditions.BROKEN_TYPE_KEY;
 import static com.sigmundgranaas.forgero.core.type.Type.*;
 
 public class StateWriter implements Writer {
@@ -56,15 +60,23 @@ public class StateWriter implements Writer {
         if (this.state instanceof Composite composite) {
             new CompositeWriter(composite).write(tooltip, context);
         }
-        var active = state.stream().getActiveProperties().toList();
+
         var passive = state.stream().getPassiveProperties().toList();
 
-        if (active.size() > 0 || passive.size() > 0) {
+        if (passive.size() > 0) {
             MutableText attributes = Text.literal(" ").append(Text.translatable(Writer.toTranslationKey("properties"))).append(": ").formatted(Formatting.GRAY);
             tooltip.add(attributes);
 
             writePassives(tooltip, context);
             writeActive(tooltip, context);
+
+            if (UnbreakableHandler.isUnbreakable(state)) {
+                tooltip.add(Text.translatable("forgero.item.unbreakable"));
+            }
+            var key = PropertyTargetCacheKey.of(state, BROKEN_TYPE_KEY);
+            if (ContainsFeatureCache.check(key)) {
+                tooltip.add(Text.translatable("forgero.item.broken"));
+            }
         }
     }
 
@@ -76,7 +88,6 @@ public class StateWriter implements Writer {
 
     private void writeActive(List<Text> tooltip, TooltipContext context) {
         var writer = new ActiveWriter();
-        state.stream().getActiveProperties().forEach(writer::addPassive);
         writer.write(tooltip, context);
     }